I could only describe this as bashing my face against a wall and expecting a different result every time.
I get this is a wrapper around c++; but how is there so little information on basic things? Suppose it is a valve product.

Why is it so hard to figure out how to properly position an item in the hand.

Why is it so hard to parent to a hand immediately? Start/Awake seem to be too early; so I have to do it in update?

Why can't I disable a hand's ability to grab?